@@ -14,6 +14,7 @@
   },
   "dependencies": {
     "@next-auth/prisma-adapter": "^1.0.5",
+    "@next/third-parties": "^14.1.3",
     "@prisma/client": "^4.16.2",
     "@types/react": "18.0.28",
     "@types/react-dom": "18.0.11",
@@ -22,7 +23,7 @@
     "date-fns": "^2.29.3",
     "eslint": "8.36.0",
     "eslint-config-next": "13.2.4",
-    "next": "13.2.4",
+    "next": "^13.5.6",
     "next-auth": "^4.20.1",
     "next-cloudinary": "^4.0.1",
     "next-nprogress-bar": "^2.1.2",
