@@ -1,8 +1,7 @@
-/** @type {import('next').NextConfig} */
+/**
+ * @type {import('next').NextConfig}
+ */
 const nextConfig = {
-    experimental: {
-        appDir: true,
-    },
     images: {
         domains: [
             "res.cloudinary.com",
@@ -11,4 +10,4 @@ const nextConfig = {
         ],
     },
 };
-module.exports = nextConfig;
+export default nextConfig;
